rename: app 폴더 밖으로 폴더 이동
